sync
git-svn-id: svn://svn.code.sf.net/p/irbem/code/trunk@116 80555d2d-b7e3-4ae1-aded-73e86dfc2014

+++ matlab/onera_desp_lib_fortran_version.m
@@ -26,5 +26,5 @@ onera_desp_lib_load;
 
 vPtr = libpointer('int32Ptr',-1);
 calllib('onera_desp_lib','irbem_fortran_version1_',vPtr);
-V = vPtr.value;
+V = double(vPtr.value);
