+++ matlab/onera_desp_lib_ntime_max.m
@@ -25,5 +25,5 @@ onera_desp_lib_load;
 
 nPtr = libpointer('int32Ptr',-1);
 calllib('onera_desp_lib','get_irbem_ntime_max1_',nPtr);
-ntime_max = nPtr.value;
+ntime_max = double(nPtr.value);
